Fix package names and move classes
